feat(admin): Sync-Seite für Admins — Campai + LADV
Closes #143
