+++ server/api/admin/sync-members.post.ts
@@ -1,7 +1,7 @@
-import { requireSuperuser } from '~~/server/utils/auth'
+import { requireAdmin } from '~~/server/utils/auth'
 import { runSyncMembers } from '~~/server/utils/sync-members'
 
 export default defineEventHandler(async (event) => {
-  await requireSuperuser(event)
+  await requireAdmin(event)
   return await runSyncMembers()
 })
